+++ src/main/java/pyororin/cryptcat/service/impl/SkipTradeServiceImpl.java
@@ -28,7 +28,7 @@ public class SkipTradeServiceImpl implements TradeService {
     public BigDecimal buy(Pair pair, OrderRequest orderRequest) {
         /* 市場最終価格(ticker.last or ticker.ask) * 注文量(amount) = 注文価格 */
         var tickerResponse = repository.retrieveTicker(CoinCheckRequest.builder().pair(pair).build());
-        var marketBuyAmount = BigDecimal.valueOf(tickerResponse.getFairBuyPrice())
+        var marketBuyAmount = tickerResponse.getFairBuyPrice()
                 .multiply(apiConfig.getAmount()).multiply(orderRequest.getRatio());
         log.info("{} {} {} {} {} {}",
                 value("kind", "exchange-skip"),
@@ -44,7 +44,7 @@ public class SkipTradeServiceImpl implements TradeService {
     public BigDecimal sell(Pair pair, OrderRequest orderRequest) {
         /* 市場最終価格(ticker.last or ticker.ask) * 注文量(amount) = 注文価格 */
         var tickerResponse = repository.retrieveTicker(CoinCheckRequest.builder().pair(pair).build());
-        var marketSellAmount = BigDecimal.valueOf(tickerResponse.getFairSellPrice())
+        var marketSellAmount = tickerResponse.getFairSellPrice()
                 .multiply(apiConfig.getAmount()).multiply(orderRequest.getRatio());
         log.info("{} {} {} {} {} {}",
                 value("kind", "exchange-skip"),
